requirements.txt
camelot_py==0.9.0
pandas==2.2.3
python-dotenv==1.0.1
